main/requirements.txt
flask==3.1.1
openai==1.78.1
python-dotenv==1.1.0